refactor: [M3-6346] - MUI v5 Migration – Components > HighlightedMarkdown
