+++ packages/manager/src/components/HighlightedMarkdown/HighlightedMarkdown.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { renderWithTheme } from 'src/utilities/testHelpers';
-import HighlightedMarkdown from './HighlightedMarkdown';
+import { HighlightedMarkdown } from './HighlightedMarkdown';
 
 const sampleMarkdown =
   '# Some markdown \n ```\n const x = function() { return true; }\n```';
