+++ packages/manager/src/components/HighlightedMarkdown/HighlightedMarkdown.tsx
@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-import classNames from 'classnames';
 import * as hljs from 'highlight.js/lib/core';
 import * as React from 'react';
-import { makeStyles } from '@mui/styles';
+import { makeStyles } from 'tss-react/mui';
 import { Theme } from '@mui/material/styles';
 import Typography from 'src/components/core/Typography';
 import 'src/formatted-text.css';
@@ -21,7 +20,7 @@ hljs.registerLanguage('javascript', javascript);
 hljs.registerLanguage('nginx', nginx);
 hljs.registerLanguage('yaml', yaml);
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles()((theme: Theme) => ({
   root: {
     '& .hljs': {
       color: theme.color.offBlack,
@@ -46,7 +45,7 @@ export interface HighlightedMarkdownProps {
 }
 
 export const HighlightedMarkdown = (props: HighlightedMarkdownProps) => {
-  const classes = useStyles();
+  const { classes, cx } = useStyles();
   const { className, language, textOrMarkdown, sanitizeOptions } = props;
   const rootRef = React.useRef<HTMLDivElement>(null);
 
@@ -85,7 +84,7 @@ export const HighlightedMarkdown = (props: HighlightedMarkdownProps) => {
 
   return (
     <Typography
-      className={classNames(classes.root, 'formatted-text', className)}
+      className={cx(classes.root, 'formatted-text', className)}
       ref={rootRef}
       dangerouslySetInnerHTML={{
         __html: sanitizedHtml,
@@ -93,5 +92,3 @@ export const HighlightedMarkdown = (props: HighlightedMarkdownProps) => {
     />
   );
 };
-
-export default HighlightedMarkdown;
