--- packages/manager/src/components/HighlightedMarkdown/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './HighlightedMarkdown';
